Code monitoring: move to enterprise
