+++ client/web/src/nav/NavLinks.tsx
@@ -22,7 +22,7 @@ import {
 import { isErrorLike } from '../../../shared/src/util/errors'
 import { Settings } from '../schema/settings.schema'
 import { InsightsNavItem } from '../insights/InsightsNavLink'
-import { CodeMonitoringNavItem } from '../code-monitoring/CodeMonitoringNavItem'
+import { CodeMonitoringNavItem } from '../enterprise/code-monitoring/CodeMonitoringNavItem'
 import { AuthenticatedUser } from '../auth'
 import { TelemetryProps } from '../../../shared/src/telemetry/telemetryService'
 import { ExtensionsNavItem } from '../extensions/ExtensionsNavItem'
